requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
requests==2.25.1
werkzeug==0.16.1
